inheritance for Contact: Phone/Email/Address
